move methods from Aimbot -> ZombiesAgent
