+++ src/main/java/dev/cigarette/module/zombies/Aimbot.java
@@ -35,13 +35,10 @@ public class Aimbot extends TickModule<ToggleWidget, Boolean> {
     private final ToggleWidget silentAim = new ToggleWidget(Text.literal("Silent Aim"), Text.literal("Doesn't snap your camera client-side.")).withDefaultState(true);
     private final ToggleWidget autoShoot = new ToggleWidget(Text.literal("Auto Shoot"), Text.literal("Automatically shoots zombies")).withDefaultState(true);
     private final ToggleWidget autoWeaponSwitch = new ToggleWidget(Text.literal("Auto Weapon Switch"), Text.literal("Automatically switch weapons")).withDefaultState(true);
-    private final ToggleWidget predictiveAim = new ToggleWidget(Text.literal("Predictive Aim"), Text.literal("Predict zombie movement for better accuracy")).withDefaultState(true);
-    private final SliderWidget predictionTicks = new SliderWidget(Text.literal("Prediction Ticks"), Text.literal("How many ticks ahead to predict zombie movement")).withBounds(1, 10, 20).withAccuracy(0);
+    public final ToggleWidget predictiveAim = new ToggleWidget(Text.literal("Predictive Aim"), Text.literal("Predict zombie movement for better accuracy")).withDefaultState(true);
+    public final SliderWidget predictionTicks = new SliderWidget(Text.literal("Prediction Ticks"), Text.literal("How many ticks ahead to predict zombie movement")).withBounds(1, 10, 20).withAccuracy(0);
 
     private KeyBinding rightClickKey = null;
-
-    private final Map<UUID, Vec3d> lastPositions = new HashMap<>();
-    private final Map<UUID, Vec3d> velocities = new HashMap<>();
 
     public Aimbot() {
         super(ToggleWidget::module, MODULE_ID, MODULE_NAME, MODULE_TOOLTIP);
@@ -51,101 +48,6 @@ public class Aimbot extends TickModule<ToggleWidget, Boolean> {
         autoWeaponSwitch.registerConfigKey("zombies.aimbot.autoWeaponSwitch");
         predictiveAim.registerConfigKey("zombies.aimbot.predictiveAim");
         predictionTicks.registerConfigKey("zombies.aimbot.predictionTicks");
-    }
-
-    @Nullable
-    private ZombiesAgent.ZombieTarget getBestTarget(ClientPlayerEntity player) {
-        HashSet<ZombiesAgent.ZombieTarget> zombies = ZombiesAgent.getZombies();
-        if (zombies.isEmpty()) {
-            return null;
-        }
-
-        ZombiesAgent.ZombieTarget bestTarget = null;
-        double minTimeToPlayer = Double.MAX_VALUE;
-
-        Vec3d playerPos = player.getPos();
-
-        for (ZombiesAgent.ZombieTarget zombie : zombies) {
-            Vec3d zombiePos = zombie.entity.getPos();
-            double distance = playerPos.distanceTo(zombiePos);
-
-            Vec3d velocity = velocities.getOrDefault(zombie.uuid, Vec3d.ZERO);
-            Vec3d playerToZombie = zombiePos.subtract(playerPos).normalize();
-            double speedTowardsPlayer = -velocity.dotProduct(playerToZombie) * 20;
-
-            if (speedTowardsPlayer <= 0.1) {
-                // Use distance as a fallback metric for stationary or retreating targets
-                if (bestTarget == null || distance < playerPos.distanceTo(bestTarget.entity.getPos())) {
-                    bestTarget = zombie;
-                }
-                continue;
-            }
-
-            double timeToPlayer = distance / speedTowardsPlayer;
-
-            if (timeToPlayer < minTimeToPlayer) {
-                minTimeToPlayer = timeToPlayer;
-                bestTarget = zombie;
-            }
-        }
-
-        return bestTarget;
-    }
-
-    private Vec3d calculatePredictedPosition(ZombiesAgent.ZombieTarget zombie, ClientPlayerEntity player) {
-        if (!predictiveAim.getRawState()) {
-            return zombie.getEndVec();
-        }
-
-        Vec3d currentPos = zombie.entity.getPos();
-        Vec3d playerPos = player.getEyePos();
-        Vec3d currentVelocity = velocities.getOrDefault(zombie.uuid, Vec3d.ZERO);
-        Vec3d pathDirection = getPathfindingDirection(zombie, currentPos, playerPos, currentVelocity);
-
-        if (pathDirection.lengthSquared() < 1.0E-7D) {
-            return zombie.getEndVec();
-        }
-
-        double distance = playerPos.distanceTo(currentPos);
-        double projectileSpeed = getProjectileSpeed();
-        double timeToTarget = (projectileSpeed > 0) ? distance / projectileSpeed : 0;
-
-        int maxPredictionTicks = predictionTicks.getRawState().intValue();
-        double maxPredictionTime = maxPredictionTicks / 20.0;
-        timeToTarget = Math.min(timeToTarget, maxPredictionTime);
-
-        Vec3d predictedBodyPos = currentPos.add(pathDirection.multiply(timeToTarget));
-
-        return predictedBodyPos.add(0, zombie.entity.getEyeHeight(zombie.entity.getPose()), 0);
-    }
-
-    private Vec3d getPathfindingDirection(ZombiesAgent.ZombieTarget zombie, Vec3d zombiePos, Vec3d playerPos, Vec3d currentVelocity) {
-        Vec3d directPath = playerPos.subtract(zombiePos).normalize();
-
-        if (currentVelocity.lengthSquared() > 1.0E-7D) {
-            Vec3d normalizedVelocity = currentVelocity.normalize();
-            double directness = normalizedVelocity.dotProduct(directPath);
-            directness = Math.max(0, Math.min(1, directness));
-
-            return directPath.multiply(1.0 - directness).add(normalizedVelocity.multiply(directness)).normalize().multiply(currentVelocity.length());
-        }
-
-        return directPath.multiply(estimateZombieSpeed(zombie) / 20.0);
-    }
-
-    private double estimateZombieSpeed(ZombiesAgent.ZombieTarget zombie) {
-        return switch (zombie.type) {
-            case ZOMBIE, SKELETON, CREEPER, WITCH -> 5.0; // ~0.25 B/t * 20 t/s
-            case BLAZE -> 8.0;
-            case WOLF -> 7.0;
-            case MAGMACUBE, SLIME -> 4.0;
-            case ENDERMITE, SILVERFISH -> 6.0;
-            default -> 5.0;
-        };
-    }
-
-    private double getProjectileSpeed() {
-        return 20.0;
     }
 
     @Override
@@ -198,40 +100,6 @@ public class Aimbot extends TickModule<ToggleWidget, Boolean> {
         cleanupTrackingData();
     }
 
-    private void updateAllZombieVelocities() {
-        HashSet<ZombiesAgent.ZombieTarget> zombies = ZombiesAgent.getZombies();
-        if (zombies.isEmpty()) {
-            velocities.clear();
-            return;
-        }
-
-        for (ZombiesAgent.ZombieTarget zombie : zombies) {
-            UUID zombieId = zombie.uuid;
-            Vec3d currentPos = zombie.entity.getPos();
-            Vec3d lastPos = lastPositions.get(zombieId);
-
-            if (lastPos != null) {
-                velocities.put(zombieId, currentPos.subtract(lastPos));
-            }
-            lastPositions.put(zombieId, currentPos);
-        }
-    }
-
-    private void cleanupTrackingData() {
-        Set<UUID> currentZombies = new HashSet<>();
-        for (ZombiesAgent.ZombieTarget zombie : ZombiesAgent.getZombies()) {
-            currentZombies.add(zombie.uuid);
-        }
-
-        lastPositions.keySet().retainAll(currentZombies);
-        velocities.keySet().retainAll(currentZombies);
-    }
-
-    @Override
-    protected void onDisabledTick(MinecraftClient client) {
-        lastPositions.clear();
-        velocities.clear();
-    }
 
     @Override
     public boolean inValidGame() {
